Moving acceptance tester into steps and pages
